+++ components/bridge/connection/sample/android/src/androidMain/kotlin/com/flipperdevices/bridge/connection/ConnectionTestApplication.kt
@@ -3,6 +3,7 @@ package com.flipperdevices.bridge.connection
 import android.app.Application
 import com.flipperdevices.bridge.connection.di.AppComponent
 import com.flipperdevices.bridge.connection.di.DaggerMergedAndroidAppComponent
+import com.flipperdevices.core.activityholder.CurrentActivityHolder
 import com.flipperdevices.core.di.ApplicationParams
 import com.flipperdevices.core.di.ComponentHolder
 import timber.log.Timber
@@ -27,5 +28,7 @@ class ConnectionTestApplication : Application() {
         ComponentHolder.components += appComponent
 
         Timber.plant(Timber.DebugTree())
+
+        CurrentActivityHolder.register(this)
     }
 }