+++ components/bridge/connection/sample/shared/src/commonMain/kotlin/com/flipperdevices/bridge/connection/screens/search/ConnectionSearchViewModel.kt
@@ -2,14 +2,25 @@ package com.flipperdevices.bridge.connection.screens.search
 
 import com.flipperdevices.bridge.connection.config.api.FDevicePersistedStorage
 import com.flipperdevices.core.ui.lifecycle.DecomposeViewModel
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.coroutines.flow.StateFlow
+import kotlinx.collections.immutable.persistentListOf
+import kotlinx.collections.immutable.toImmutableList
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-abstract class ConnectionSearchViewModel(
-    private val persistedStorage: FDevicePersistedStorage
+class ConnectionSearchViewModel @Inject constructor(
+    private val persistedStorage: FDevicePersistedStorage,
+    searchDelegatesFactories: MutableSet<ConnectionSearchDelegate.Factory>
 ) : DecomposeViewModel() {
-    abstract fun getDevicesFlow(): StateFlow<ImmutableList<ConnectionSearchItem>>
+    private val searchDelegates = searchDelegatesFactories.map { it(viewModelScope) }
+    private val combinedFlow = combine(searchDelegates.map { it.getDevicesFlow() }) { flows ->
+        flows.toList().flatten().toImmutableList()
+    }.stateIn(viewModelScope, SharingStarted.Lazily, persistentListOf())
+
+    fun getDevicesFlow() = combinedFlow
+
     fun onDeviceClick(searchItem: ConnectionSearchItem) {
         viewModelScope.launch {
             if (searchItem.isAdded) {