+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 # 1.8.2 - In Progress
 
+- [Feature] Add USB connection option to sample
 - [Feature] Migrate app onto new BLE api
 - [Fix] Fix CI for desktop sample app
-
 
 # 1.8.1